requirements.txt
python-dotenv>=1.0
ollama>=0.4
requests>=2.31
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
adjustText>=0.8
